rename(#4): dto 위치 수정
