+++ service/payment/server/src/main/java/com/sparta/payment/presentation/dto/PaymentResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.payment.application.dto;
+package com.sparta.payment.presentation.dto;
 
 import com.sparta.payment.domain.entity.PaymentState;
 import java.time.LocalDateTime;
